split Scoreboard into three components (Score,ScoreBoard,Form), and moved all the relevant code-base under these components. Issue #6
