+++ src/components/Score.js
@@ -1,0 +1,17 @@
+const Score = (props) => {
+    return (
+        <div className='row'>
+            <div className='row-sub' id="index">
+                {props.index}
+            </div>
+            <div className='row-sub' id="name">
+                {props.user}
+            </div>
+            <div className='row-sub' id="score">
+                {props.score}
+            </div>
+        </div>
+    )
+};
+
+export default Score;
